requirements.txt
annotated-types==0.5.0
anyio==3.7.1
click==8.1.4
fastapi==0.100.0
fuzzywuzzy==0.18.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
joblib==1.3.1
numpy==1.25.1
pandas==2.0.3
pydantic==2.0.2
pydantic_core==2.1.2
pyodbc==4.0.39
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
scikit-learn==1.3.0
scipy==1.11.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.18
starlette==0.27.0
threadpoolctl==3.1.0
typing_extensions==4.7.1
tzdata==2023.3
uvicorn==0.22.0
